% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod_calculate_age.R
\name{cod_calculate_age}
\alias{cod_calculate_age}
\alias{cod_calculate_ages}
\title{Calculate age at death based on date of birth and date of death}
\usage{
cod_calculate_age(dob, dod, date_format = "\%Y-\%m-\%d", codedit = TRUE)

cod_calculate_ages(dob, dod, date_format = "\%Y-\%m-\%d", codedit = TRUE)
}
\arguments{
\item{dob}{Date of birth. This should ideally be in standard ISO extended
date format of \emph{"YYYY-MM-DD"} as specified in the default value for
\code{date_format}.}

\item{dod}{Date of death. This should ideally be in standard ISO extended
date format of \emph{YYYY-MM_DD"} as specified in the default value for
\code{date_format}.}

\item{date_format}{Format for date values provided. Date formatting is
handled using \code{\link[=strptime]{strptime()}} hence this needs to be specified based on what
\code{\link[=strptime]{strptime()}} requires for its \code{format} argument. By default, this is set
to the standard ISO extended date format expressed as \emph{"\%Y-\%m-\%d"} which
corresponds to \emph{"YYYY-MM-DD"}.}

\item{codedit}{Logical. Should output be based on the CoDEdit version 2
coding rules. Default to TRUE.}
}
\value{
Values for age in days, months and years. IF \code{codedit} is TRUE,
a tibble with age value and age type as required by CoDEdit.
}
\description{
Calculate age at death based on date of birth and date of death
}
\examples{
cod_calculate_age("1977-11-05", Sys.Date())
cod_calculate_age("1965-05-20", "2023-10-03")

}
