% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod_check_age.R
\name{cod_check_age}
\alias{cod_check_age}
\title{Check age values in cause of death data based on CoDEdit rules}
\usage{
cod_check_age(age_value, age_type, age_type_code = c("D", "M", "Y"))
}
\arguments{
\item{age_value}{An integer value or vector of values for age based on the
CoDEdit rules.}

\item{age_type}{A vector of values for age type based on the CoDEdit rules.
This should either be "D" for age in days, "M" for age
in months, or "Y" for age in years. If values are different from these,
then \code{age_type_code} should be specified to correspond to the day, month,
and year values of \code{age_type}.}

\item{age_type_code}{A character or integer vector of 3 values that indicate
which values are to be considered pertaining to days (first value in the
vector), to months (second value in the vector), or years (third value
in the vector).}
}
\value{
A tibble with number of rows equal to length of \code{age_value} and
two columns for age_check_score and age_check_note.
}
\description{
Check age values in cause of death data based on CoDEdit rules
}
\examples{
cod_check_age(120, "Y")
cod_check_age(28, "D")
cod_check_age(32, "D")

}
