% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod_check_code.R
\name{cod_check_code}
\alias{cod_check_code}
\alias{cod_check_code_icd10}
\alias{cod_check_code_icd11}
\title{Check cause of death code for code entry mistakes and/or code completeness}
\usage{
cod_check_code(cod, version = c("icd10", "icd11"))

cod_check_code_icd10(cod)

cod_check_code_icd11(cod)
}
\arguments{
\item{cod}{A character value or vector of values for cause of death code/s.}

\item{version}{A character value for ICD version used. This should be either
\emph{"icd10"} or \emph{"icd11"}. Default is \emph{"icd10"}.}
}
\value{
A tibble with 2 columns/fields. First is an integer value indicating
whether there is an issue with the cause of death code provided in relation
to a potential code entry mistake and/or and issue of code completeness.
}
\description{
Check cause of death code for code entry mistakes and/or code completeness
}
\examples{
cod_check_code("U100")
cod_check_code("2C6Z", version = "icd11")

}
