% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod_summary.R
\name{cod_check_code_summary}
\alias{cod_check_code_summary}
\title{Summarise cause-of-death check results}
\usage{
cod_check_code_summary(cod_check, simplify = FALSE)
}
\arguments{
\item{cod_check}{A data.frame output of the various \verb{cod_check_code_*}
functions}

\item{simplify}{Logical. Should output be converted into a data.frame?
Default is FALSE.}
}
\value{
If \code{simplify} is FALSE (default), a list of summary check outputs.
Otherwise, a tabulated summary of check outputs.
}
\description{
Summarise cause-of-death check results
}
\examples{
cod_check_code(
  cod_data_raw_example$code, version = "icd11",
  sex = cod_data_raw_example$sex, age = cod_data_raw_example$age
) |>
  cod_check_code_summary()

}
