% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod_check_input.R
\name{cod_check_codedit_input}
\alias{cod_check_codedit_input}
\title{Check structure and values of input data to CoDEdit tool}
\usage{
cod_check_codedit_input(df)
}
\arguments{
\item{df}{A data.frame with 6 columns with names \code{"FreeId"}, \code{"Sex"},
\code{"Age Value"}, \code{"Age Type"}, \code{"Code"}, and \code{"Death Date"} and compatible
with the input data required by the CoDEdit tool.}
}
\value{
A data.frame containing check codes and check notes for each row
and variable identified with the \code{FreeId} of \code{df}.
}
\description{
Check structure and values of input data to CoDEdit tool
}
\examples{
cod_check_codedit_input(icd10_example)

}
