% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod_check_sex.R
\name{cod_check_sex}
\alias{cod_check_sex}
\title{Check sex values in cause of death data based on CoDEdit rules}
\usage{
cod_check_sex(sex_value, sex_code = c(1, 2))
}
\arguments{
\item{sex_value}{An integer value or vector of values for age based on the
CoDEdit rules.}

\item{sex_code}{A character or integer vector of 2 values that indicate
which values are to be considered pertaining to males (first value in the
vector) or to females (second value in the vector). Default is 1 for male
and 2 for female.}
}
\value{
A tibble with number of rows equal to length of \code{sex_value} and
two columns for sex_check and sex_check_note.
}
\description{
Check sex values in cause of death data based on CoDEdit rules
}
\examples{
cod_check_sex("m", c("m", "f"))
cod_check_sex("male", c("male", "female"))
cod_check_sex(1, 1:2)

}
