% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod_data_raw_example}
\alias{cod_data_raw_example}
\title{Example raw cause of death dataset}
\format{
A data frame with 6 columns and 20 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{id} \tab Unique identifier \cr
   \emph{sex} \tab Sex of deceased \cr
   \emph{age} \tab Age of diseased in years \cr
   \emph{code} \tab ICD 11 cause of death code \cr
   \emph{dod} \tab Date of death \cr
   \emph{dob} \tab Date of birth \cr
}
}
\usage{
cod_data_raw_example
}
\description{
Example raw cause of death dataset
}
\examples{
cod_data_raw_example

}
\keyword{datasets}
