% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod_recode_age_type.R
\name{cod_recode_age_type}
\alias{cod_recode_age_type}
\title{Recode age type of cause-of-death data based on CoDEdit rules}
\usage{
cod_recode_age_type(age_type, age_type_code = c("D", "M", "Y"))
}
\arguments{
\item{age_type}{A vector of values for age type based on the CoDEdit rules.
This should either be "D" for age in days, "M" for age
in months, or "Y" for age in years. If values are different from these,
then \code{age_type_code} should be specified to correspond to the day, month,
and year values of \code{age_type}.}

\item{age_type_code}{A character or integer vector of 3 values that indicate
which values are to be considered pertaining to days (first value in the
vector), to months (second value in the vector), or years (third value
in the vector).}
}
\value{
A character value or vector of values containing either "D", "M",
or "Y" for \emph{days}, \emph{months}, or \emph{years} respectively.
}
\description{
Recode age type of cause-of-death data based on CoDEdit rules
}
\examples{
cod_recode_age_type(
  age_type = c(rep("d", 3), rep("m", 2), rep("y", 3)),
  age_type_code = c("d", "m", "y")
)

}
