% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod_recode_sex.R
\name{cod_recode_sex}
\alias{cod_recode_sex}
\title{Recode cause of death data on sex}
\usage{
cod_recode_sex(sex_value, sex_code = c(1, 2), codedit = TRUE)
}
\arguments{
\item{sex_value}{A character or integer value or vector of values signifying
the sex.}

\item{sex_code}{A character or integer vector of 2 values that indicate
which values are to be considered pertaining to males (first value in the
vector) or to females (second value in the vector).}

\item{codedit}{Logical. Should output be based on the CoDEdit version 2
coding rules. Default to TRUE.}
}
\value{
An integer value or vector of values containing either 1 for males
or 2 for females. If \code{codedit = TRUE}, values not equal to the \code{sex_code}
values are coded as 9 (integer). Otherwise, it is coded as NA_integer_.
}
\description{
Recode cause of death data on sex
}
\examples{
cod_recode_sex(
  sex_value = c(rep("m", 2), rep("f", 3)),
  sex_code = c("m", "f")
)

}
