% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod_structure_input.R
\name{cod_structure_input}
\alias{cod_structure_input}
\title{Structure cause-of-death data into CoDEdit tool input data}
\usage{
cod_structure_input(df, sex, sex_code = c(1, 2), dob, dod, code, id = NULL)
}
\arguments{
\item{df}{A data.frame of raw cause-of-death data with the following
required variables that contains values for sex, date of birth,
date of death, and cause-of-death code.}

\item{sex}{A character value for the variable name in \code{df} containing
the values for sex.}

\item{sex_code}{A character or integer vector of 2 values that indicate
which values are to be considered pertaining to males (first value in the
vector) or to females (second value in the vector).}

\item{dob}{A character value for the variable name in \code{df} containing
the values for date of birth.}

\item{dod}{A character value for the variable name in \code{df} containing
the values for date of death.}

\item{code}{A character value for the variable name in \code{df} containing
the values for cause-of-death code.}

\item{id}{A character value for the variable name in \code{df} containing unique
record identifiers. Default to NULL. If NULL, unique record identifiers
will be generated.}
}
\value{
A data.frame with 6 columns and number of rows equal to \code{df} with
names \code{"FreeId"}, \code{"Sex"}, \code{"Age Value"}, \code{"Age Type"}, \code{"Code"}, and
\code{"Death Date"}.
}
\description{
Structure cause-of-death data into CoDEdit tool input data
}
\examples{
df <- data.frame(
  id = 1:3,
  sex = c(1, 1, 2),
  dob = c("1977-11-05", "1971-04-04", "2012-08-13"),
  dod = c("2024-06-28", "2023-10-11", "2023-09-25"),
  code = c("P219", "O230", "Q913")
)

cod_structure_input(df, sex = "sex", dob = "dob", dod = "dod", code = "code")

}
