% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeditr-package.R
\docType{package}
\name{codeditr}
\alias{codeditr-package}
\alias{codeditr}
\title{Implementing Cause-of-Death Data Checks Based on the WHO CoDEdit Tool}
\description{
The World Health Organization's CoDEdit electronic tool is intended to help
producers of cause-of-death statistics in strengthening their capacity to
perform routine checks on their data. This package ports the original tool
built using Microsoft Access into R so as to leverage the utility and
function of the original tool into a usable application program interface
that can be used for building more universal tools or for creating
programmatic scientific workflows aimed at routine, automated, and
large-scale monitoring of cause-of-death data quality.
}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Anita Makori [copyright holder]
}

Other contributors:
\itemize{
  \item Sanjeev Pugazhendhi [thesis advisor, reviewer]
  \item Williams Udoh Ituen-Umanah [reviewer]
  \item Proochista Ariana [reviewer, copyright holder]
}

}
\keyword{internal}
