% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_icd10_code_range}
\alias{expand_icd10_code_range}
\title{Enumerate ICD 10 codes given a code range}
\usage{
expand_icd10_code_range(code_range)
}
\arguments{
\item{code_range}{A character value or a vector of character values
indicating a range of ICD 10 codes. See Details for syntax of code range/s.}
}
\value{
A vector of ICD 10 codes that are within the range of codes
specified by \code{code_range}.
}
\description{
Enumerate ICD 10 codes given a code range
}
\examples{
expand_icd10_code_range("A71.0-A71.9")
expand_icd10_code_range("F50.1,F50.3-F50.9")

}
