% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_age_values}
\alias{get_age_values}
\title{Get various age values for all three age types}
\usage{
get_age_values(age_value, age_type = c("D", "M", "Y"))
}
\arguments{
\item{age_value}{An integer value for the age}

\item{age_type}{The age type of the specified age value. Can be either
\emph{"D"} for day, \emph{"M"} for month, or \emph{"Y"} for year.}
}
\value{
A named list of age values in days, months, and years.
}
\description{
Get various age values for all three age types
}
\examples{
get_age_values(1, "Y")

}
