% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10_cod_by_sex}
\alias{icd10_cod_by_sex}
\title{Sex-specific causes of death for ICD 10}
\format{
A data frame with 3 columns and 880 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{code} \tab ICD 10 Cause of Death code \cr
   \emph{title} \tab Cause of death title \cr
   \emph{sex} \tab Sex - 1 for male; 2 for female \cr
}
}
\source{
https://icd.who.int/browse10/Content/statichtml/ICD10Volume2_en_2019.pdf
}
\usage{
icd10_cod_by_sex
}
\description{
Sex-specific causes of death for ICD 10
}
\examples{
icd10_cod_by_sex

}
\keyword{datasets}
