% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10_cod_child}
\alias{icd10_cod_child}
\title{Child-specific cause-of-death for ICD 10}
\format{
A data frame with 2 columns and 122 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{code} \tab ICD 10 Cause of Death code \cr
   \emph{title} \tab Cause of death title \cr
}
}
\source{
https://www.icd10data.com/ICD10CM/Codes/Rules/Pediatric_Codes
}
\usage{
icd10_cod_child
}
\description{
Child-specific cause-of-death for ICD 10
}
\examples{
icd10_cod_child

}
\keyword{datasets}
