% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10_example}
\alias{icd10_example}
\title{Example death records dataset with ICD10 cause-of-death coding}
\format{
A data frame with 6 columns and 3613 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{FreeId} \tab Record identifier \cr
   \emph{Sex} \tab Sex: 1 = Male; 2 = Female; 9 = unknown \cr
   \emph{Age Value} \tab Integer value for age \cr
   \emph{Age Type} \tab Is the age value in days (D), months (M), or years (Y) \cr
   \emph{Code} \tab ICD10 code for cause-of-death \cr
   \emph{Death Date} \tab Date of death in year format \cr
}
}
\source{
https://www.who.int/standards/classifications/classification-of-diseases/services/codedit-tool
}
\usage{
icd10_example
}
\description{
Example death records dataset with ICD10 cause-of-death coding
}
\examples{
icd10_example

}
\keyword{datasets}
