% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10_unlikely_cod}
\alias{icd10_unlikely_cod}
\title{Unlikely causes of death for ICD 10}
\format{
A data frame with 2 columns and 424 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{code} \tab ICD 10 Cause of Death code \cr
   \emph{title} \tab Cause of death title \cr
}
}
\source{
https://icd.who.int/browse10/Content/statichtml/ICD10Volume2_en_2019.pdf
}
\usage{
icd10_unlikely_cod
}
\description{
Unlikely causes of death for ICD 10
}
\examples{
icd10_unlikely_cod

}
\keyword{datasets}
