% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd11_cod_by_sex}
\alias{icd11_cod_by_sex}
\title{Sex-specific causes of death for ICD 11}
\format{
A data frame with 3 columns and 547 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{code} \tab ICD 11 Cause of Death code \cr
   \emph{title} \tab Cause of death title \cr
   \emph{gender} \tab Gender - male or female \cr
}
}
\source{
https://icdcdn.who.int/icd11referenceguide/en/html/index.html#list-of-categories-limited-to-or-more-likely-to-occur-in-female-persons
}
\usage{
icd11_cod_by_sex
}
\description{
Sex-specific causes of death for ICD 11
}
\examples{
icd11_cod_by_sex

}
\keyword{datasets}
