% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd11_cod_child}
\alias{icd11_cod_child}
\title{Child-specific cause of death for ICD 11}
\format{
A data frame with 2 columns and 149 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{code} \tab ICD 11 Cause of Death code \cr
   \emph{title} \tab Cause of death title \cr
}
}
\usage{
icd11_cod_child
}
\description{
Child-specific cause of death for ICD 11
}
\examples{
icd11_cod_child


}
\keyword{datasets}
