% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd11_unlikely_cod}
\alias{icd11_unlikely_cod}
\title{Unlikely caused of death for ICD 11}
\format{
A data frame with 2 columns and 269 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{code} \tab ICD 11 Cause of Death code \cr
   \emph{title} \tab Cause of death title \cr
}
}
\source{
https://icd.who.int/valuesets/viewer/582/en
}
\usage{
icd11_unlikely_cod
}
\description{
Unlikely caused of death for ICD 11
}
\examples{
icd11_unlikely_cod

}
\keyword{datasets}
