% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_ill_defined_icd11}
\alias{list_ill_defined_icd11}
\title{List ill-defined ICD 11 codes}
\usage{
list_ill_defined_icd11()

list_ill_defined_icd11()
}
\value{
An character vector of ICD 11 codes classified as ill-defined for
cause of death
}
\description{
Based on https://icdcdn.who.int/icd11referenceguide/en/html/index.html#list-of-illdefined-conditions
}
\examples{
list_ill_defined_icd11()

}
