% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod_check_sex.R
\name{cod_check_sex}
\alias{cod_check_sex}
\title{Check sex values in cause of death data based on CoDEdit rules}
\usage{
cod_check_sex(sex_value, sex_code)
}
\arguments{
\item{sex_value}{An integer value or vector of values for age based on the
CoDEdit rules.}

\item{sex_code}{A character or integer vector of 2 values that indicate
which values are to be considered pertaining to males (first value in the
vector) or to females (second value in the vector).}
}
\description{
Check sex values in cause of death data based on CoDEdit rules
}
