% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod_check_sex.R
\name{cod_check_sex}
\alias{cod_check_sex}
\title{Check sex values in cause of death data based on CoDEdit rules}
\usage{
cod_check_sex(sex_value)
}
\arguments{
\item{sex_value}{An integer value or vector of values for age based on the
CoDEdit rules.}
}
\description{
Check sex values in cause of death data based on CoDEdit rules
}
