% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_ill_defined_icd11}
\alias{list_ill_defined_icd11}
\title{List ill-defined ICD 11 codes}
\usage{
list_ill_defined_icd11()
}
\value{
An character vector of ICD 11 codes classified as ill-defined for
cause-of-death
}
\description{
List ill-defined ICD 11 codes
}
\examples{
list_ill_defined_icd11()

}
